Use volume attributes for REST clients
This commit allows for the use of the information in an incoming
VolumeCreate request to initialize the REST clients to the Gluster
servers. Since DeleteVolume() does not receive any volume attributes, we
would have needed to fetch all PVs and search them for one that
corresponds to the volume being deleted. To avoid this, we establish a
cache of all previously connected GlusterFS servers which is updated on
a VolumeCreate() for a non-existing volume and on a cache miss.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>