Containerize build process
With this change, the entire build process runs as a part of building
the container image. To accomplish this, the existing Dockerfile was
changed to be multi-stage. The resulting final image is similar in size
and layer count to the old build process.

Along with moving to multi-stage:
- The golang build now targets a static executable.
- The Gluster client dependencies are pulled from the CentOS storage sig
repository to get the latest stable GLuster version instead of the EOL
version that is a part of base CentOS 7.5.
- The builder container uses the recommended upstream for OpenShift
golang projects

Signed-off-by: John Strunk <jstrunk@redhat.com>